% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deriv12}
\alias{deriv12}
\title{Compute 1st and 2nd derivative}
\usage{
deriv12(fun, x, delta = 1e-04, fx = NULL, ...)
}
\arguments{
\item{fun}{function for which to compute the gradient.}

\item{x}{numeric vector of values at which to compute the
gradient.}

\item{delta}{the amount to subtract and add to \code{x} when computing the
central difference.}

\item{fx}{optional value of \code{fun(x, ...)}.}

\item{...}{additional arguments to \code{fun}.}
}
\value{
a list with components
\item{gradient}{the first derivative vector}
\item{Hessian}{the second derivative matrix}
}
\description{
This function computes the 1st and 2nd derivative (gradient and
Hessian) of a function
with vector-valued input and scalar output using a central finite
difference method.
}
\details{
The function has to return a single scalar numerical value and take
as argument a numeric vector of the same length as \code{x}. The function also
has to be evaluable in a neighborhood around \code{x}, i.e. at
\code{x - delta} and at \code{x + delta} for all elements in \code{x}.
}
\examples{
## Make a function:
f <- function(x) sum(dnorm(x))

## compute gradient and Hessian (first two derivatives):
D12 <- deriv12(fun=f, x=1:2)

## Compare with analytical Hessian:
df <- deriv3(~ dnorm(x1) + dnorm(x2), c("x1", "x2"), func=TRUE)
dg <- function(x1, x2) as.vector(attributes(df(x1, x2))[[1]])
dH <- function(x1, x2)
    matrix(unlist(attributes(df(x1, x2))[[2]]), ncol=2)

dg(1, 2) - D12$gradient
dH(1, 2) - D12$Hessian

stopifnot(
    isTRUE(all.equal(dg(1, 2), D12$gradient, tol=1e-6)),
    isTRUE(all.equal(dH(1, 2), D12$Hessian, tol=1e-6))
    )
}
\author{
Rune Haubo Bojesen Christensen
}
\seealso{
Other derivatives: \code{\link{gradient}};
  \code{\link{hessian}}
}
\keyword{utilities}

