% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drop.coef}
\alias{drop.coef}
\title{Ensure full rank design matrix}
\usage{
drop.coef(X, silent = FALSE)
}
\arguments{
\item{X}{a design matrix, e.g., the result of
\code{\link{model.matrix}} possibly of less than full column rank,
i.e., with redundant parameters. Works for \code{ncol(X) >= 0} and
\code{nrow(X) >= 0}.}

\item{silent}{should a message not be issued if X is column rank
deficient?}
}
\value{
The design matrix \code{X} without redundant columns.
}
\description{
Coefficients (columns) are dropped from a design matrix to
ensure that it has full rank.
}
\details{
Redundant columns of the design matrix are identified with the
 LINPACK implementation of the \code{\link{qr}} decomposition and
 removed. The returned design matrix will have \code{qr(X)$rank}
 columns.
}
\examples{

## Make artificial data:
df <- data.frame(g = factor(letters[rep(1:2, c(4, 2))]),
                 f = factor(LETTERS[c(1,2,1,2,1,1)]))


## Make design matrix:
X <- model.matrix(~ g * f, data=df)
## Compare ncol before and after dropping redundant columns:
X
newX <- drop.coef(X)
newX
## Optionally suppres message (e.g. inside functions):
newX <- drop.coef(X, silent=TRUE)

## Essentially this is being computed:
qr.X <- qr(X, tol = 1e-7, LAPACK = FALSE)
newX <- X[, qr.X$pivot[1:qr.X$rank], drop = FALSE]
## is newX of full column rank?
ncol(newX) == qr(newX)$rank
## the number of columns being dropped:
ncol(X) - ncol(newX)

## A 'real' example:
## if(require(ordinal)) {
##     X <- model.matrix( ~ PRODID * DAY, data = soup)
##     ncol(X)
##     newX <- drop.coef(X)
##     ncol(newX)
## }
}
\author{
Rune Haubo Bojesen Christensen
}
\seealso{
\code{\link{qr}} and \code{\link{lm}}
}
\keyword{models}

